h3: introduce Event::PriorityUpdate
This commit adds a new Event::PriorityUpdate that is returned by the
`poll()` method. It only occurs on servers because only clients can send
PRIORITY_UPDATE frames.

Event:PriorityUpdate behaves like other edge-triggered events. The first
time a server receives a PRIORITY_UPDATE frame, the event is fired and
subsequent `poll()` will not generate more events. Application are
required to use the new `recv_priority_update()` method to read all
received PRIORITY_UPDATE frames before the trigger is rearmed.

PRIORITY_UPDATE frames are sent on the control stream and include a
reference to request stream IDs. Due to reordering, it is possible that
a PRIORITY_UPDATE frame can be received before the HEADERS of the
request. Because there are no ordering guarantees at the wire, the
ordering of events is unimportant. This commit chooses to present
Event::PriorityUpdate ahead of other events.

Clients might send multiple PRIORITY_UPDATE frames between server reads,
but only the most up to date value matters. To accommodate this, only
the single most recent PRIORITY_UPDATE is stored in quiche until it is
read via `recv_priority_update()`.

Although, PRIORITY_UPDATE frames can reference streams unknown at the
server (see above). There are limits to the streams that they refer to.
A PRIORITY_UPDATE frame that exceeds the server's maximum bidi stream ID
is treated as an error, as defined in the spec. A PRIORITY_UPDATE that
refers to a collected stream is completely ignored and will not cause an
Event::PriorityUpdate to be generated. In order to apply these
restrictions, the quiche library has had a minor change to expose the
required information about streams.